More variable cleanup
