+++ .gitignore
@@ -2,7 +2,6 @@
 explainaboard/example/test-re.tsv
 explainaboard/example/test-re.tsv
 explainaboard/example/test_re.tsv
-explainaboard/tasks/re/errorAnalysis.py
 explainaboard/example/test-re.tsv
 explainaboard/example/test-re.tsv
 explainaboard/tasks/re/test_re.tsv
