+++ explainaboard/tasks/absa/eval_spec.py
@@ -2,7 +2,7 @@
 import explainaboard.error_analysis as ea
 import explainaboard.data_utils as du
 import numpy
-
+from collections import OrderedDict
 
 def process_all(file_path, size_of_bin=10, dataset='atis', model='lstm-self-attention'):
     """
@@ -29,7 +29,6 @@ def process_all(file_path, size_of_bin=10, dataset='atis', model='lstm-self-atte
     ece :the ece of this file
     dic :the details of the ECE information in json format
     """
-    from collections import OrderedDict
 
     probability_list, right_or_not_list = du.get_probability_right_or_not(file_path)
 
@@ -133,6 +132,14 @@ def file_to_list(path_file):
     fin.close()
     return sent1_list, sent2_list, true_label_list, pred_label_list
 
+def get_error_case(aspect_list, sent_list, true_label_list, pred_label_list):
+    error_case_list = []
+    for aspect, sent, true_label, pred_label in zip(aspect_list, sent_list, true_label_list, pred_label_list):
+        if true_label != pred_label:
+            error_case_list.append(
+                true_label + "|||" + pred_label + "|||" + ea.format4json2(aspect) + "|||" + ea.format4json2(sent))
+    return error_case_list
+
 def evaluate(task_type="ner", analysis_type="single", systems=[], output="./output.json", is_print_ci=False,
              is_print_case=False, is_print_ece=False):
     path_text = ""
@@ -166,7 +173,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
 
     error_case_list = []
     if is_print_case:
-        error_case_list = ea.get_error_case_absa(aspect_list, sent_list, true_label_list, pred_label_list)
+        error_case_list = get_error_case(aspect_list, sent_list, true_label_list, pred_label_list)
         print(" -*-*-*- the number of error casse:\t", len(error_case_list))
 
     # Confidence Interval of Holistic Performance
@@ -185,7 +192,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
     print("------------------ Holistic Result----------------------")
     print(holistic_performance)
 
-    def __selectBucktingFunc(func_name, func_setting, dict_obj):
+    def __select_bucketing_func(func_name, func_setting, dict_obj):
         if func_name == "bucket_attribute_SpecifiedBucketInterval":
             return ea.bucket_attribute_specified_bucket_interval(dict_obj, eval(func_setting))
         elif func_name == "bucket_attribute_SpecifiedBucketValue":
@@ -208,15 +215,15 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
 
     for aspect, func in dict_aspect_func.items():
         # print(aspect, dict_span2aspect_val[aspect])
-        dict_bucket2span[aspect] = __selectBucktingFunc(func[0], func[1], dict_span2aspect_val[aspect])
+        dict_bucket2span[aspect] = __select_bucketing_func(func[0], func[1], dict_span2aspect_val[aspect])
         # print(aspect, dict_bucket2span[aspect])
         # exit()
         dict_bucket2span_pred[aspect] = ea.bucket_attribute_specified_bucket_interval(dict_span2aspect_val_pred[aspect],
                                                                                       dict_bucket2span[aspect].keys())
-        # dict_bucket2span_pred[aspect] = __selectBucktingFunc(func[0], func[1], dict_span2aspect_val_pred[aspect])
-        dict_bucket2f1[aspect] = ea.get_bucket_acc_with_error_case_absa(dict_bucket2span[aspect],
-                                                                     dict_bucket2span_pred[aspect], dict_sid2sample,
-                                                                     is_print_ci, is_print_case)
+        # dict_bucket2span_pred[aspect] = __select_bucketing_func(func[0], func[1], dict_span2aspect_val_pred[aspect])
+        dict_bucket2f1[aspect] = get_bucket_acc_with_error_case(dict_bucket2span[aspect],
+                                                                dict_bucket2span_pred[aspect], dict_sid2sample,
+                                                                is_print_ci, is_print_case)
         aspect_names.append(aspect)
     print("aspect_names: ", aspect_names)
 
@@ -238,7 +245,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
         print(k + ":\t" + str(v))
     print("")
 
-    def beautifyInterval(interval):
+    def beautify_interval(interval):
 
         if type(interval[0]) == type("string"):  ### pay attention to it
             return interval[0]
@@ -252,13 +259,13 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
                 bk_name = range1_r + range1_l
                 return bk_name
 
-    dict_fineGrained = {}
+    dict_fine_grained = {}
     for aspect, metadata in dict_bucket2f1.items():
-        dict_fineGrained[aspect] = []
+        dict_fine_grained[aspect] = []
         for bucket_name, v in metadata.items():
             # print("---------debug--bucket name old---")
             # print(bucket_name)
-            bucket_name = beautifyInterval(bucket_name)
+            bucket_name = beautify_interval(bucket_name)
             # print("---------debug--bucket name new---")
             # print(bucket_name)
 
@@ -268,15 +275,15 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
             confidence_low = format(v[2], '.4g')
             confidence_up = format(v[3], '.4g')
 
-            # for saving errorlist -- finegrained version
+            # for saving errorlist -- fine_grained version
             bucket_error_case = v[4]
 
             # instantiation
-            dict_fineGrained[aspect].append({"bucket_name": bucket_name, "bucket_value": bucket_value, "num": n_sample,
+            dict_fine_grained[aspect].append({"bucket_name": bucket_name, "bucket_value": bucket_value, "num": n_sample,
                                              "confidence_low": confidence_low, "confidence_up": confidence_up,
                                              "bucket_error_case": bucket_error_case})
 
-    # dict_fineGrained[aspect].append({"bucket_name":bucket_name, "bucket_value":bucket_value, "num":n_sample, "confidence_low":confidence_low, "confidence_up":confidence_up, "bucket_error_case":[]})
+    # dict_fine_grained[aspect].append({"bucket_name":bucket_name, "bucket_value":bucket_value, "num":n_sample, "confidence_low":confidence_low, "confidence_up":confidence_up, "bucket_error_case":[]})
 
     obj_json = ea.load_json(path_json_input)
 
@@ -289,7 +296,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
     obj_json["model"]["results"]["overall"]["performance"] = holistic_performance
     obj_json["model"]["results"]["overall"]["confidence_low"] = confidence_low
     obj_json["model"]["results"]["overall"]["confidence_up"] = confidence_up
-    obj_json["model"]["results"]["fine_grained"] = dict_fineGrained
+    obj_json["model"]["results"]["fine_grained"] = dict_fine_grained
 
     # add errorAnalysis -- holistic
     obj_json["model"]["results"]["overall"]["error_case"] = error_case_list
@@ -307,3 +314,52 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
     obj_json["data"]["output"] = path_comb_output
 
     ea.save_json(obj_json, fn_write_json)
+
+
+def get_bucket_acc_with_error_case(dict_bucket2span, dict_bucket2span_pred, dict_sid2sentpair, is_print_ci,
+                                   is_print_case):
+    # The structure of span_true or span_pred
+    # 2345|||Positive
+    # 2345 represents sentence id
+    # Positive represents the "label" of this instance
+
+    dict_bucket2f1 = {}
+
+    for bucket_interval, spans_true in dict_bucket2span.items():
+        spans_pred = []
+
+        # print('bucket_interval: ',bucket_interval)
+        if bucket_interval not in dict_bucket2span_pred.keys():
+            # print(bucket_interval)
+            raise ValueError("Predict Label Bucketing Errors")
+        else:
+            spans_pred = dict_bucket2span_pred[bucket_interval]
+
+        # loop over samples from a given bucket
+        error_case_bucket_list = []
+
+        if is_print_case:
+            for info_true, info_pred in zip(spans_true, spans_pred):
+                sid_true, label_true = info_true.split("|||")
+                sid_pred, label_pred = info_pred.split("|||")
+                if sid_true != sid_pred:
+                    continue
+
+                sent = dict_sid2sentpair[sid_true]
+                if label_true != label_pred:
+                    error_case_info = label_true + "|||" + label_pred + "|||" + sent
+                    error_case_bucket_list.append(error_case_info)
+
+        accuracy_each_bucket = ea.accuracy(spans_pred, spans_true)
+        confidence_low, confidence_up = 0, 0
+        if is_print_ci:
+            confidence_low, confidence_up = ea.compute_confidence_interval_acc(spans_pred, spans_true)
+
+        dict_bucket2f1[bucket_interval] = [accuracy_each_bucket, len(spans_true), confidence_low, confidence_up,
+                                           error_case_bucket_list]
+
+        # print(error_case_bucket_list)
+
+        print("accuracy_each_bucket:\t", accuracy_each_bucket)
+
+    return ea.sort_dict(dict_bucket2f1)
