+++ explainaboard/tasks/cws/eval_spec.py
@@ -1,6 +1,9 @@
+from random import choices
+
 import explainaboard.error_analysis as ea
 import numpy
 import pickle
+import codecs
 
 
 def read_data(corpus_type, fn, column_no=-1, delimiter=' '):
@@ -9,7 +12,7 @@ def read_data(corpus_type, fn, column_no=-1, delimiter=' '):
     tag_sequences = list()
     total_word_sequences = list()
     total_tag_sequences = list()
-    with ea.codecs.open(fn, 'r', 'utf-8') as f:
+    with codecs.open(fn, 'r', 'utf-8') as f:
         lines = f.readlines()
     curr_words = list()
     curr_tags = list()
@@ -42,14 +45,14 @@ def read_data(corpus_type, fn, column_no=-1, delimiter=' '):
     return total_word_sequences, total_tag_sequences, word_sequences, tag_sequences
 
 
-def get_aspect_value(test_word_sequences, test_trueTag_sequences, test_word_sequences_sent,
-                   test_trueTag_sequences_sent, dict_precomputed_path, dict_aspect_func):
-    def getSententialValue(test_trueTag_sequences_sent, test_word_sequences_sent):
+def get_aspect_value(test_word_sequences, test_true_tag_sequences, test_word_sequences_sent,
+                   test_true_tag_sequences_sent, dict_precomputed_path, dict_aspect_func):
+    def get_sentential_value(test_true_tag_sequences_sent, test_word_sequences_sent):
 
         eDen = []
         sentLen = []
 
-        for i, test_sent in enumerate(test_trueTag_sequences_sent):
+        for i, test_sent in enumerate(test_true_tag_sequences_sent):
             pred_chunks = set(ea.get_chunks(test_sent))
 
             num_entityToken = 0
@@ -80,13 +83,12 @@ def get_aspect_value(test_word_sequences, test_trueTag_sequences, test_word_sequ
     for aspect, fun in dict_aspect_func.items():
         dict_span2aspect_val[aspect] = {}
 
-    eDen_list, sentLen_list = [], []
-    eDen_list, sentLen_list = getSententialValue(test_trueTag_sequences_sent,
-                                                 test_word_sequences_sent)
+    eDen_list, sentLen_list = get_sentential_value(test_true_tag_sequences_sent,
+                                                   test_word_sequences_sent)
 
     dict_pos2sid = ea.get_pos2sentid(test_word_sequences_sent)
     dict_ap2rp = ea.get_token_position(test_word_sequences_sent)
-    all_chunks = ea.get_chunks(test_trueTag_sequences)
+    all_chunks = ea.get_chunks(test_true_tag_sequences)
 
     dict_span2sid = {}
     dict_chunkid2span = {}
@@ -103,7 +105,7 @@ def get_aspect_value(test_word_sequences, test_trueTag_sequences, test_word_sequ
         # print(span_cnt.encode("utf-8").decode("utf-8"))
         span_cnt = span_cnt.encode("gbk", "ignore").decode("gbk", "ignore")
         # print(sys.getdefaultencoding())
-        span_type = ''.join(test_trueTag_sequences[idx_start:idx_end])
+        span_type = ''.join(test_true_tag_sequences[idx_start:idx_end])
 
         span_pos = str(idx_start) + "|||" + str(idx_end) + "|||" + span_type
 
@@ -200,11 +202,11 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
 
     confidence_low_overall, confidence_up_overall = 0, 0
     if is_print_ci:
-        confidence_low_overall, confidence_up_overall = ea.compute_confidence_interval_f1_cws(dict_span2sid.keys(),
-                                                                                              dict_span2sid_pred.keys(),
-                                                                                              dict_span2sid,
-                                                                                              dict_span2sid_pred,
-                                                                                              n_times=10)
+        confidence_low_overall, confidence_up_overall = compute_confidence_interval_f1(dict_span2sid.keys(),
+                                                                                       dict_span2sid_pred.keys(),
+                                                                                       dict_span2sid,
+                                                                                       dict_span2sid_pred,
+                                                                                       n_times=10)
 
     print("confidence_low_overall:\t", confidence_low_overall)
     print("confidence_up_overall:\t", confidence_up_overall)
@@ -214,7 +216,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
 
     # print(f1(list_true_tags_token, list_pred_tags_token)["f1"])
 
-    def __selectBucktingFunc(func_name, func_setting, dict_obj):
+    def __select_bucketing_func(func_name, func_setting, dict_obj):
         if func_name == "bucket_attribute_SpecifiedBucketInterval":
             return ea.bucket_attribute_specified_bucket_interval(dict_obj, eval(func_setting))
         elif func_name == "bucket_attribute_SpecifiedBucketValue":
@@ -240,16 +242,16 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
 
     for aspect, func in dict_aspect_func.items():
         # print(aspect, dict_span2aspect_val[aspect])
-        dict_bucket2span[aspect] = __selectBucktingFunc(func[0], func[1], dict_span2aspect_val[aspect])
+        dict_bucket2span[aspect] = __select_bucketing_func(func[0], func[1], dict_span2aspect_val[aspect])
         # print(aspect, dict_bucket2span[aspect])
         # exit()
         dict_bucket2span_pred[aspect] = ea.bucket_attribute_specified_bucket_interval(dict_span2aspect_val_pred[aspect],
                                                                                       dict_bucket2span[aspect].keys())
-        dict_bucket2f1[aspect], error_case_list = ea.getBucketF1_cws(dict_bucket2span[aspect],
-                                                                    dict_bucket2span_pred[aspect], dict_span2sid,
-                                                                    dict_span2sid_pred, dict_chunkid2span,
-                                                                    dict_chunkid2span_pred, list_true_tags_token,
-                                                                    list_pred_tags_token, is_print_ci, is_print_case)
+        dict_bucket2f1[aspect], error_case_list = get_bucket_f1(dict_bucket2span[aspect],
+                                                                dict_bucket2span_pred[aspect], dict_span2sid,
+                                                                dict_span2sid_pred, dict_chunkid2span,
+                                                                dict_chunkid2span_pred, list_true_tags_token,
+                                                                list_pred_tags_token, is_print_ci, is_print_case)
         aspect_names.append(aspect)
     print("aspect_names: ", aspect_names)
 
@@ -272,7 +274,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
         print(k + ":\t" + str(v))
     print("")
 
-    def beautifyInterval(interval):
+    def beautify_interval(interval):
 
         if type(interval[0]) == type("string"):  ### pay attention to it
             return interval[0]
@@ -286,13 +288,13 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
                 bk_name = range1_r + range1_l
                 return bk_name
 
-    dict_fineGrained = {}
+    dict_fine_grained = {}
     for aspect, metadata in dict_bucket2f1.items():
-        dict_fineGrained[aspect] = []
+        dict_fine_grained[aspect] = []
         for bucket_name, v in metadata.items():
             # print("---------debug--bucket name old---")
             # print(bucket_name)
-            bucket_name = beautifyInterval(bucket_name)
+            bucket_name = beautify_interval(bucket_name)
             # print("---------debug--bucket name new---")
             # print(bucket_name)
 
@@ -305,7 +307,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
             error_entity_list = v[4]
 
             # instantiation
-            dict_fineGrained[aspect].append({"bucket_name": bucket_name, "bucket_value": bucket_value, "num": n_sample,
+            dict_fine_grained[aspect].append({"bucket_name": bucket_name, "bucket_value": bucket_value, "num": n_sample,
                                              "confidence_low": confidence_low, "confidence_up": confidence_up,
                                              "bucket_error_case": error_entity_list[
                                                                   0:int(len(error_entity_list) / 10)]})
@@ -321,9 +323,202 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
     obj_json["model"]["results"]["overall"]["performance"] = holistic_performance
     obj_json["model"]["results"]["overall"]["confidence_low"] = confidence_low_overall
     obj_json["model"]["results"]["overall"]["confidence_up"] = confidence_up_overall
-    obj_json["model"]["results"]["fine_grained"] = dict_fineGrained
+    obj_json["model"]["results"]["fine_grained"] = dict_fine_grained
 
     # Save error cases: overall
     obj_json["model"]["results"]["overall"]["error_case"] = error_case_list[0:int(len(error_case_list) / 10)]
 
     ea.save_json(obj_json, fn_write_json)
+
+
+def compute_confidence_interval_f1(spans_true, spans_pred, dict_span2sid, dict_span2sid_pred, n_times=1000):
+    n_data = len(dict_span2sid)
+    sample_rate = ea.get_sample_rate(n_data)
+    n_sampling = int(n_data * sample_rate)
+    print("sample_rate:\t", sample_rate)
+    print("n_sampling:\t", n_sampling)
+
+    dict_sid2span_salient = {}
+    for span in spans_true:
+        # print(span)
+        if len(span.split("|||")) != 3:
+            break
+        sid = dict_span2sid[span]
+        if sid in dict_sid2span_salient.keys():
+            dict_sid2span_salient[sid].append(span)
+        else:
+            dict_sid2span_salient[sid] = [span]
+
+    dict_sid2span_salient_pred = {}
+    for span in spans_pred:
+        sid = dict_span2sid_pred[span]
+        if sid in dict_sid2span_salient_pred.keys():
+            dict_sid2span_salient_pred[sid].append(span)
+        else:
+            dict_sid2span_salient_pred[sid] = [span]
+
+    performance_list = []
+    confidence_low, confidence_up = 0, 0
+    for i in range(n_times):
+        sample_index_list = choices(range(n_data), k=n_sampling)
+
+        true_label_bootstrap_list = []
+        pred_label_bootstrap_list = []
+        for ind, sid in enumerate(sample_index_list):
+
+            if sid in dict_sid2span_salient.keys():
+                true_label_list = dict_sid2span_salient[sid]
+                true_label_list_revised = [true_label + "|||" + str(ind) for true_label in true_label_list]
+                true_label_bootstrap_list += true_label_list_revised
+
+            if sid in dict_sid2span_salient_pred.keys():
+                pred_label_list = dict_sid2span_salient_pred[sid]
+                pred_label_list_revised = [pred_label + "|||" + str(ind) for pred_label in pred_label_list]
+                pred_label_bootstrap_list += pred_label_list_revised
+
+        f1, p, r = ea.evaluate_chunk_level(pred_label_bootstrap_list, true_label_bootstrap_list)
+        performance_list.append(f1)
+
+    if n_times != 1000:
+        confidence_low, confidence_up = ea.mean_confidence_interval(performance_list)
+    else:
+        performance_list.sort()
+        confidence_low = performance_list[24]
+        confidence_up = performance_list[974]
+
+    # print("\n")
+    # print("confidence_low:\t", confidence_low)
+    # print("confidence_up:\t", confidence_up)
+
+    return confidence_low, confidence_up
+
+
+def get_error_case(dict_pos2tag, dict_pos2tag_pred, dict_chunkid2span_sent, dict_chunkid2span_sent_pred,
+                   list_true_tags_token, list_pred_tags_token):
+    error_case_list = []
+    for pos, tag in dict_pos2tag.items():
+
+        true_label = tag
+        pred_label = ""
+        # print(dict_chunkid2span_sent.keys())
+        if pos + "|||" + tag not in dict_chunkid2span_sent.keys():
+            continue
+        span_sentence = dict_chunkid2span_sent[pos + "|||" + tag]
+
+        if pos in dict_pos2tag_pred.keys():
+            pred_label = dict_pos2tag_pred[pos]
+            if true_label == pred_label:
+                continue
+        # print(pos + "\t" + true_label + "\t" + pred_label)
+        else:
+            start = int(pos.split("|||")[0])
+            end = int(pos.split("|||")[1])
+            pred_label = "".join(list_pred_tags_token[start:end])
+        # print(pred_label)
+        error_case = span_sentence + "|||" + true_label + "|||" + pred_label
+        error_case_list.append(error_case)
+
+    for pos, tag in dict_pos2tag_pred.items():
+
+        true_label = ""
+        pred_label = tag
+        if pos + "|||" + tag not in dict_chunkid2span_sent_pred.keys():
+            continue
+        span_sentence = dict_chunkid2span_sent_pred[pos + "|||" + tag]
+
+        if pos in dict_pos2tag.keys():
+            true_label = dict_pos2tag[pos]
+            if true_label == pred_label:
+                continue
+        else:
+            start = int(pos.split("|||")[0])
+            end = int(pos.split("|||")[1])
+            true_label = "".join(list_true_tags_token[start:end])
+        error_case = span_sentence + "|||" + true_label + "|||" + pred_label
+        error_case_list.append(error_case)
+
+    # for v in error_case_list:
+    # 	print(len(error_case_list))
+    # 	print(v)
+    # print(error_case_list)
+
+    return error_case_list
+
+
+def get_bucket_f1(dict_bucket2span, dict_bucket2span_pred, dict_span2sid, dict_span2sid_pred, dict_chunkid2span,
+                  dict_chunkid2span_pred, list_true_tags_token, list_pred_tags_token, is_print_ci, is_print_case):
+    dict_bucket2f1 = {}
+
+    # predict:  2_3 -> NER
+    dict_pos2tag_pred = {}
+    if is_print_case:
+        for k_bucket_eval, spans_pred in dict_bucket2span_pred.items():
+            for span_pred in spans_pred:
+                pos_pred = "|||".join(span_pred.split("|||")[0:2])
+                tag_pred = span_pred.split("|||")[-1]
+                dict_pos2tag_pred[pos_pred] = tag_pred
+
+    # true:  2_3 -> NER
+    dict_pos2tag = {}
+    if is_print_case:
+        for k_bucket_eval, spans in dict_bucket2span.items():
+            for span in spans:
+                pos = "|||".join(span.split("|||")[0:2])
+                tag = span.split("|||")[-1]
+                dict_pos2tag[pos] = tag
+
+    error_case_list = []
+    if is_print_case:
+        error_case_list = get_error_case(dict_pos2tag, dict_pos2tag_pred, dict_chunkid2span, dict_chunkid2span_pred,
+                                         list_true_tags_token, list_pred_tags_token)
+
+    # print(len(error_case_list))
+    # print(error_case_list)
+
+    for bucket_interval, spans_true in dict_bucket2span.items():
+        spans_pred = []
+        if bucket_interval not in dict_bucket2span_pred.keys():
+            raise ValueError("Predict Label Bucketing Errors")
+        else:
+            spans_pred = dict_bucket2span_pred[bucket_interval]
+
+        confidence_low, confidence_up = 0, 0
+        if is_print_ci:
+            confidence_low, confidence_up = compute_confidence_interval_f1(spans_true, spans_pred, dict_span2sid,
+                                                                           dict_span2sid_pred)
+
+        confidence_low = format(confidence_low, '.3g')
+        confidence_up = format(confidence_up, '.3g')
+
+        f1, p, r = ea.evaluate_chunk_level(spans_pred, spans_true)
+
+        error_entity_list = []
+        if is_print_case:
+            for span_true in spans_true:
+                if span_true not in spans_pred:
+                    # print(span_true)
+                    pos_true = "|||".join(span_true.split("|||")[0:2])
+                    tag_true = span_true.split("|||")[-1]
+
+                    if pos_true in dict_pos2tag_pred.keys():
+                        tag_pred = dict_pos2tag_pred[pos_true]
+                        if tag_pred != tag_true:
+                            error_entity_list.append(
+                                dict_chunkid2span[span_true] + "|||" + tag_true + "|||" + dict_pos2tag_pred[pos_true])
+                    # print(dict_chunkid2span[span_true] + "|||" + tag_true + "|||" + dict_pos2tag_pred[pos_true])
+                    else:
+                        start = int(pos_true.split("|||")[0])
+                        end = int(pos_true.split("|||")[1])
+                        pred_label = "".join(list_pred_tags_token[start:end])
+                        error_entity_list.append(dict_chunkid2span[span_true] + "|||" + tag_true + "|||" + pred_label)
+
+            # print(dict_chunkid2span[span_true] + "|||" + tag_true + "|||" + pred_label)
+
+        dict_bucket2f1[bucket_interval] = [f1, len(spans_true), confidence_low, confidence_up, error_entity_list]
+
+    # if bucket_interval[0] == 1.0:
+    # 	print("debug-f1:",f1)
+    # 	print(spans_pred[0:20])
+    # 	print(spans_true[0:20])
+    # print("dict_bucket2f1: ",dict_bucket2f1)
+    return ea.sort_dict(dict_bucket2f1), error_case_list
