+++ explainaboard/tasks/nli/eval_spec.py
@@ -162,11 +162,11 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
             dict_precomputed_path[aspect] = path_precomputed + "_" + aspect + ".pkl"
             print("precomputed directory:\t", dict_precomputed_path[aspect])
 
-    sent1_list, sent2_list, true_label_list, pred_label_list = ea.file_to_list_nli(path_text)
+    sent1_list, sent2_list, true_label_list, pred_label_list = file_to_list(path_text)
 
     error_case_list = []
     if is_print_case:
-        error_case_list = ea.get_error_case_nli(sent1_list, sent2_list, true_label_list, pred_label_list)
+        error_case_list = get_error_case(sent1_list, sent2_list, true_label_list, pred_label_list)
         print(" -*-*-*- the number of error casse:\t", len(error_case_list))
 
     # Confidence Interval of Holistic Performance
@@ -187,7 +187,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
 
     # print(f1(list_true_tags_token, list_pred_tags_token)["f1"])
 
-    def __selectBucktingFunc(func_name, func_setting, dict_obj):
+    def __select_bucketing_func(func_name, func_setting, dict_obj):
         if func_name == "bucket_attribute_SpecifiedBucketInterval":
             return ea.bucket_attribute_specified_bucket_interval(dict_obj, eval(func_setting))
         elif func_name == "bucket_attribute_SpecifiedBucketValue":
@@ -212,15 +212,15 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
 
     for aspect, func in dict_aspect_func.items():
         # print(aspect, dict_span2aspect_val[aspect])
-        dict_bucket2span[aspect] = __selectBucktingFunc(func[0], func[1], dict_span2aspect_val[aspect])
+        dict_bucket2span[aspect] = __select_bucketing_func(func[0], func[1], dict_span2aspect_val[aspect])
         # print(aspect, dict_bucket2span[aspect])
         # exit()
         dict_bucket2span_pred[aspect] = ea.bucket_attribute_specified_bucket_interval(dict_span2aspect_val_pred[aspect],
                                                                                       dict_bucket2span[aspect].keys())
-        # dict_bucket2span_pred[aspect] = __selectBucktingFunc(func[0], func[1], dict_span2aspect_val_pred[aspect])
-        dict_bucket2f1[aspect] = ea.get_bucket_acc_with_error_case_nli(dict_bucket2span[aspect],
-                                                                    dict_bucket2span_pred[aspect], dict_sid2sentpair,
-                                                                    is_print_ci, is_print_case)
+        # dict_bucket2span_pred[aspect] = __select_bucketing_func(func[0], func[1], dict_span2aspect_val_pred[aspect])
+        dict_bucket2f1[aspect] = get_bucket_acc_with_error_case(dict_bucket2span[aspect],
+                                                                dict_bucket2span_pred[aspect], dict_sid2sentpair,
+                                                                is_print_ci, is_print_case)
         aspect_names.append(aspect)
     print("aspect_names: ", aspect_names)
 
@@ -240,7 +240,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
         print(k + ":\t" + str(v))
     print("")
 
-    def beautifyInterval(interval):
+    def beautify_interval(interval):
 
         if type(interval[0]) == type("string"):  ### pay attention to it
             return interval[0]
@@ -254,13 +254,13 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
                 bk_name = range1_r + range1_l
                 return bk_name
 
-    dict_fineGrained = {}
+    dict_fine_grained = {}
     for aspect, metadata in dict_bucket2f1.items():
-        dict_fineGrained[aspect] = []
+        dict_fine_grained[aspect] = []
         for bucket_name, v in metadata.items():
             # print("---------debug--bucket name old---")
             # print(bucket_name)
-            bucket_name = beautifyInterval(bucket_name)
+            bucket_name = beautify_interval(bucket_name)
             # print("---------debug--bucket name new---")
             # print(bucket_name)
 
@@ -270,11 +270,11 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
             confidence_low = format(v[2], '.4g')
             confidence_up = format(v[3], '.4g')
 
-            # for saving errorlist -- finegrained version
+            # for saving errorlist -- fine_grained version
             bucket_error_case = v[4]
 
             # instantiation
-            dict_fineGrained[aspect].append({"bucket_name": bucket_name, "bucket_value": bucket_value, "num": n_sample,
+            dict_fine_grained[aspect].append({"bucket_name": bucket_name, "bucket_value": bucket_value, "num": n_sample,
                                              "confidence_low": confidence_low, "confidence_up": confidence_up,
                                              "bucket_error_case": bucket_error_case})
 
@@ -289,7 +289,7 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
     obj_json["model"]["results"]["overall"]["performance"] = holistic_performance
     obj_json["model"]["results"]["overall"]["confidence_low"] = confidence_low
     obj_json["model"]["results"]["overall"]["confidence_up"] = confidence_up
-    obj_json["model"]["results"]["fine_grained"] = dict_fineGrained
+    obj_json["model"]["results"]["fine_grained"] = dict_fine_grained
 
     # add errorAnalysis -- holistic
     obj_json["model"]["results"]["overall"]["error_case"] = error_case_list
@@ -304,3 +304,73 @@ def evaluate(task_type="ner", analysis_type="single", systems=[], output="./outp
     obj_json["model"]["results"]["calibration"] = dic_calibration
 
     ea.save_json(obj_json, fn_write_json)
+
+
+def get_bucket_acc_with_error_case(dict_bucket2span, dict_bucket2span_pred, dict_sid2sentpair, is_print_ci,
+                                   is_print_case):
+    # The structure of span_true or span_pred
+    # 2345|||Positive
+    # 2345 represents sentence id
+    # Positive represents the "label" of this instance
+
+    dict_bucket2f1 = {}
+
+    for bucket_interval, spans_true in dict_bucket2span.items():
+        spans_pred = []
+        if bucket_interval not in dict_bucket2span_pred.keys():
+            raise ValueError("Predict Label Bucketing Errors")
+        else:
+            spans_pred = dict_bucket2span_pred[bucket_interval]
+
+        # loop over samples from a given bucket
+        error_case_bucket_list = []
+        if is_print_case:
+            for info_true, info_pred in zip(spans_true, spans_pred):
+                sid_true, label_true = info_true.split("|||")
+                sid_pred, label_pred = info_pred.split("|||")
+                if sid_true != sid_pred:
+                    continue
+
+                sent = dict_sid2sentpair[sid_true]
+                if label_true != label_pred:
+                    error_case_info = label_true + "|||" + label_pred + "|||" + sent
+                    error_case_bucket_list.append(error_case_info)
+
+        accuracy_each_bucket = ea.accuracy(spans_pred, spans_true)
+        confidence_low, confidence_up = 0, 0
+        if is_print_ci:
+            confidence_low, confidence_up = ea.compute_confidence_interval_acc(spans_pred, spans_true)
+        dict_bucket2f1[bucket_interval] = [accuracy_each_bucket, len(spans_true), confidence_low, confidence_up,
+                                           error_case_bucket_list]
+
+    return ea.sort_dict(dict_bucket2f1)
+
+
+def get_error_case(sent1_list, sent2_list, true_label_list, pred_label_list):
+    error_case_list = []
+    for sent1, sent2, true_label, pred_label in zip(sent1_list, sent2_list, true_label_list, pred_label_list):
+        if true_label != pred_label:
+            error_case_list.append(
+                true_label + "|||" + pred_label + "|||" + ea.format4json2(sent1) + "|||" + ea.format4json2(sent2))
+    return error_case_list
+
+
+def file_to_list(path_file):
+    sent1_list = []
+    sent2_list = []
+    true_label_list = []
+    pred_label_list = []
+    fin = open(path_file, "r")
+    for line in fin:
+        line = line.rstrip("\n")
+        if len(line.split("\t")) < 4:
+            continue
+        sent1, sent2, true_label, pred_label = line.split("\t")[0], line.split("\t")[1], line.split("\t")[2], \
+                                               line.split("\t")[3]
+        sent1_list.append(sent1)
+        sent2_list.append(sent2)
+        true_label_list.append(true_label)
+        pred_label_list.append(pred_label)
+
+    fin.close()
+    return sent1_list, sent2_list, true_label_list, pred_label_list
